+++ lib/src/tags/for_loop.dart
@@ -8,10 +8,16 @@ class ForLoopTag implements Tag {
   String get name => "forLoop";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies)
+  {
     // 'var' is a required attribute
     final varName = attributes["var"];
-    if (varName == null || varName.isEmpty) throw Exception("<$name> 'var' attribute is required.");
+    if (varName == null || varName.isEmpty) {
+      throw Exception("<$name> 'var' attribute is required.");
+    }
 
     final dependenciesScope = attributes["dependenciesScope"];
     final begin = CommonUtils.tryParseInt(element.getAttribute("begin")) ?? 0;
@@ -23,7 +29,7 @@ class ForLoopTag implements Tag {
     for (var index = begin; step > 0 ? index < end : index > end; index += step) {
       final deps = XWidget.scopeDependencies(dependencies, dependenciesScope);
       deps[varName] = index;
-      final tagChildren = XWidget.inflateXmlElementChildren(element, deps);
+      final tagChildren = XWidget.inflateXmlElementChildren(element, deps, excludeText: true);
       children.addAll(tagChildren);
     }