+++ lib/src/tags/if_else.dart
@@ -4,9 +4,11 @@ import '../xwidget.dart';
 
 /// A pair of tags that conditionally render portions of the UI.
 ///
-/// If the `test` expression evaluates to `true`, then <if>'s children are rendered, otherwise <else>'s children
-/// are rendered. The `test` expression must return a bool, otherwise an [Exception] is thrown. The <else> tag
-/// must be a direct child of an <if> tag, if present, otherwise it will be ignored.
+/// If the `test` expression evaluates to `true`, then <if>'s children are
+/// rendered, otherwise <else>'s children are rendered. The `test` expression
+/// must return a bool, otherwise an [Exception] is thrown. The <else> tag
+/// must be a direct child of an <if> tag, if present, otherwise it will be
+/// ignored.
 ///
 /// <if> Attributes:
 /// - test (required): A bool expression i.e. ${user.name == 'Mike'}
@@ -17,14 +19,25 @@ class IfElseTag implements Tag {
   String get name => "if";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies)
+  {
     // 'test' is a required attribute
-    if (!attributes.containsKey("test")) throw Exception("<$name> 'test' attribute is required.");
+    if (!attributes.containsKey("test")) {
+      throw Exception("<$name> 'test' attribute is required.");
+    }
 
     final test = attributes["test"];
     final elementOrElse = _isTrue(test) ? element : _findElseElement(element);
     if (elementOrElse != null) {
-      return XWidget.inflateXmlElementChildren(elementOrElse, dependencies, excludeElements: {"else"});
+      return XWidget.inflateXmlElementChildren(
+          elementOrElse,
+          dependencies,
+          excludeElements: {"else"},
+          excludeText: true
+      );
     }
     return null;
   }